+++ src/com/cnaude/chairs/ChairWatcher.java
@@ -5,8 +5,8 @@
 package com.cnaude.chairs;
 
 import java.util.ArrayList;
-import net.minecraft.server.v1_5_R3.DataWatcher;
-import net.minecraft.server.v1_5_R3.WatchableObject;
+import net.minecraft.server.v1_6_R2.DataWatcher;
+import net.minecraft.server.v1_6_R2.WatchableObject;
 
 /**
  *
@@ -23,8 +23,7 @@ public class ChairWatcher extends DataWatcher {
     @Override
     public ArrayList<WatchableObject> b() {
         ArrayList<WatchableObject> list = new ArrayList<WatchableObject>();
-        WatchableObject wo = new WatchableObject(0, 0, Byte.valueOf(this.metadata));
-        //WatchableObject wo = new WatchableObject(0, 0, 4);
+        WatchableObject wo = new WatchableObject(0, 0, this.metadata);        
         list.add(wo);
         return list;
     }
